+++ packages/terra-modal-manager/CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Unreleased
 ----------
+### Breaking Change
+* Consumed DisclosureManager v3.0 - Updated to Context-provided consumer API vs. prop injection
+* Removed `withModalManager()` export to prevent confusion with the new context changes
+
+### Removed
+* Removed dependency on terra-app-delegate package
 
 3.14.0 - (November 15, 2018)
 ------------------